Adds binary sensors and bumps pyplaato to 0.0.8
