+++ homeassistant/components/plaato/const.py
@@ -11,4 +11,4 @@ PLACEHOLDER_DOCS_URL = "docs_url"
 PLACEHOLDER_DEVICE_TYPE = "device_type"
 PLACEHOLDER_DEVICE_NAME = "device_name"
 DOCS_URL = "https://www.home-assistant.io/integrations/plaato/"
-PLATFORMS = ["sensor"]
+PLATFORMS = ["sensor", "binary_sensor"]
