+++ homeassistant/components/plaato/manifest.json
@@ -6,5 +6,5 @@
   "dependencies": ["webhook"],
   "after_dependencies": ["cloud"],
   "codeowners": ["@JohNan"],
-  "requirements": ["pyplaato==0.0.7"]
+  "requirements": ["pyplaato==0.0.8"]
 }
