+++ requirements_all.txt
@@ -1616,7 +1616,7 @@ pypck==0.7.7
 pypjlink2==1.2.1
 
 # homeassistant.components.plaato
-pyplaato==0.0.7
+pyplaato==0.0.8
 
 # homeassistant.components.point
 pypoint==2.0.0
